lab_2/pyproject.toml
[project]
name = "lab-2"
version = "0.1.0"
description = ""
authors = [
    {name = "Konrad Springer",email = "kossarpl@gmail.com"}
]
readme = "README.md"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "notification-system"
version = "1.0.0"
description = "System zarządzania i wysyłki powiadomień"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.1"
sqlalchemy = "^2.0.23"
celery = "^5.3.4"
redis = "^5.0.1"
python-dotenv = "^1.0.0"
pydantic = "^2.5.0"
tenacity = "^8.2.3"
psycopg2-binary = "^2.9.9"
uvicorn = "^0.24.0"
python-multipart = "^0.0.6"
sendgrid = "^6.10.0"
firebase-admin = "^6.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
httpx = "^0.25.1"